refactor(/src/main/java/lotto/Lotto.java): refactor LottoReward enum을 LottoEnum 클래스로 이동
